Optimize
Optimize by using fast type ints, and use uint_fast8_t in some spots for even more performance

+++ src/CMD_Args.cpp
@@ -1,8 +1,8 @@
 #include "inc/CMD_Args.hpp"
 
-void parseArgs(int argc, char** argv, bool* debugMode) {
+void parseArgs(int_fast32_t argc, char** argv, bool* debugMode) {
 	// Check if argument is --debug-mode
-	for (int i = 1; i < argc; i++) {
+	for (uint_fast8_t i = 1; i < argc; i++) {
 		if (strcmp(argv[i], "--debug-mode") == 0) {
 			*debugMode = true;
 		}
